+++ build_tools/install_rocm_from_artifacts.py
@@ -1,26 +1,28 @@
 #!/usr/bin/env python
-"""provision_machine.py
+"""install_rocm_from_artifacts.py
 
-This script helps CI workflows, developers and testing suites easily provision TheRock to their environment.
-It provisions TheRock to an output directory from one of these sources:
+This script helps CI workflows, developers and testing suites easily install TheRock to their environment using artifacts.
+It installs TheRock to an output directory from one of these sources:
 - GitHub CI workflow run
 - GitHub release tag
 - An existing installation of TheRock
 
 Usage:
-python build_tools/provision_machine.py [--output-dir OUTPUT_DIR] [--amdgpu-family AMDGPU_FAMILY] (--run-id RUN_ID | --release RELEASE | --input-dir INPUT_DIR)
+python build_tools/install_rocm_from_artifacts.py [--output-dir OUTPUT_DIR] [--amdgpu-family AMDGPU_FAMILY] (--run-id RUN_ID | --release RELEASE | --input-dir INPUT_DIR)
+                                        [--blas | --no-blas] [--fft | --no-fft] [--miopen | --no-miopen] [--prim | --no-prim]
+                                        [--rand | --no-rand] [--rccl | --no-rccl] [--tests | --no-tests] [--base-only]
 
 Examples:
 - Downloads the gfx94X S3 artifacts from GitHub CI workflow run 14474448215 (from https://github.com/ROCm/TheRock/actions/runs/14474448215) to the default output directory `therock-build`:
-    - `python build_tools/provision_machine.py --run-id 14474448215 --amdgpu-family gfx94X-dcgpu`
-- Downloads the latest gfx110X artifacts from GitHub release tag `nightly-release` to the specified output directory `build`:
-    - `python build_tools/provision_machine.py --release latest --amdgpu-family gfx110X-dgpu --output-dir build`
+    - `python build_tools/install_rocm_from_artifacts.py --run-id 14474448215 --amdgpu-family gfx94X-dcgpu --tests`
 - Downloads the version `6.4.0rc20250416` gfx110X artifacts from GitHub release tag `nightly-release` to the specified output directory `build`:
-    - `python build_tools/provision_machine.py --release 6.4.0rc20250416 --amdgpu-family gfx110X-dgpu --output-dir build`
+    - `python build_tools/install_rocm_from_artifacts.py --release 6.4.0rc20250416 --amdgpu-family gfx110X-dgpu --output-dir build`
 - Downloads the version `6.4.0.dev0+8f6cdfc0d95845f4ca5a46de59d58894972a29a9` gfx120X artifacts from GitHub release tag `dev-release` to the default output directory `therock-build`:
-    - `python build_tools/provision_machine.py --release 6.4.0.dev0+8f6cdfc0d95845f4ca5a46de59d58894972a29a9 --amdgpu-family gfx120X-all`
+    - `python build_tools/install_rocm_from_artifacts.py --release 6.4.0.dev0+8f6cdfc0d95845f4ca5a46de59d58894972a29a9 --amdgpu-family gfx120X-all`
 
 You can select your AMD GPU family from this file https://github.com/ROCm/TheRock/blob/59c324a759e8ccdfe5a56e0ebe72a13ffbc04c1f/cmake/therock_amdgpu_targets.cmake#L44-L81
+
+By default for CI workflow retrieval, all artifacts (excluding test artifacts) will be downloaded. For specific artifacts, pass in the flag such as `--rand` (RAND artifacts) For test artifacts, pass in the flag `--tests` (test artifacts). For base artifacts only, pass in the flag `--base-only`
 
 Note: the script will overwrite the output directory argument. If no argument is passed, it will overwrite the default "therock-build" directory.
 """
@@ -31,17 +33,17 @@ from fetch_artifacts import (
     retrieve_enabled_artifacts,
     s3_bucket_exists,
 )
+import json
 import os
-from packaging.version import Version, InvalidVersion
 from pathlib import Path
 import platform
-import requests
+import re
 import shutil
 import subprocess
 import sys
 import tarfile
 from _therock_utils.artifacts import ArtifactPopulator
-from tqdm import tqdm
+from urllib.request import urlopen, Request
 
 
 def log(*args, **kwargs):
@@ -54,14 +56,9 @@ def _untar_files(output_dir, destination):
     Retrieves all tar files in the output_dir, then extracts all files to the output_dir
     """
     output_dir_path = Path(output_dir).resolve()
-    # In order to get better visibility on untar-ing status, tqdm adds a progress bar
     log(f"Extracting {destination.name} to {output_dir}")
     with tarfile.open(destination) as extracted_tar_file:
-        for member in tqdm(
-            iterable=extracted_tar_file.getmembers(),
-            total=len(extracted_tar_file.getmembers()),
-        ):
-            extracted_tar_file.extract(member=member, path=output_dir_path)
+        extracted_tar_file.extractall(output_dir_path)
     destination.unlink()
 
 
@@ -97,19 +94,20 @@ def _get_github_release_assets(release_tag, amdgpu_family, release_version):
     if gh_token:
         headers["Authentication"] = f"Bearer {gh_token}"
 
-    response = requests.get(github_release_url, headers=headers)
-    if response.status_code == 403:
-        log(
-            f"Error when retrieving GitHub release assets for release tag '{release_tag}'. This is most likely a rate limiting issue, so please try again"
-        )
-        return
-    elif response.status_code != 200:
-        log(
-            f"Error when retrieving GitHub release assets for release tag '{release_tag}'. Exiting..."
-        )
-        return
+    request = Request(github_release_url, headers=headers)
+    with urlopen(request) as response:
+        if response.status == 403:
+            log(
+                f"Error when retrieving GitHub release assets for release tag '{release_tag}'. This is most likely a rate limiting issue, so please try again"
+            )
+            return
+        elif response.status != 200:
+            log(
+                f"Error when retrieving GitHub release assets for release tag '{release_tag}' with status code {response.status}. Exiting..."
+            )
+            return
 
-    release_data = response.json()
+        release_data = json.loads(response.read().decode("utf-8"))
 
     # We retrieve the most recent release asset that matches the amdgpu_family
     # In the cases of "nightly-release" or "dev-release", this will retrieve the the specified version or latest
@@ -145,17 +143,12 @@ def _download_github_release_asset(asset_data, output_dir):
     destination = Path(output_dir) / asset_name
     headers = {"Accept": "application/octet-stream"}
     # Making the API call to retrieve the asset
-    response = requests.get(asset_url, stream=True, headers=headers)
+    request = Request(asset_url, headers=headers)
 
-    # Downloading the asset in chunks to destination
-    # In order to get better visibility on downloading status, tqdm adds a progress bar
-    total_size = int(response.headers.get("content-length", 0))
-    block_size = 1024
-    with tqdm(total=total_size, unit="B", unit_scale=True) as progress_bar:
-        with open(destination, "wb") as file:
-            for chunk in response.iter_content(block_size * block_size):
-                progress_bar.update(len(chunk))
-                file.write(chunk)
+    with urlopen(request) as response_obj, open(destination, "wb") as file:
+        # Downloading the asset to destination
+        log(f"Downloading tar file to {str(destination)}")
+        shutil.copyfileobj(response_obj, file)
 
     # After downloading the asset, untar-ing the file
     _untar_files(output_dir, destination)
@@ -173,11 +166,10 @@ def retrieve_artifacts_by_run_id(args):
         log(f"S3 artifacts for {run_id} does not exist. Exiting...")
         return
 
-    args.all = True
-
     # Retrieving base and all math-lib tar artifacts and downloading them to output_dir
     retrieve_base_artifacts(args, run_id, output_dir)
-    retrieve_enabled_artifacts(args, True, amdgpu_family, run_id, output_dir)
+    if not args.base_only:
+        retrieve_enabled_artifacts(args, amdgpu_family, run_id, output_dir)
 
     # Flattening artifacts from .tar* files then removing .tar* files
     log(f"Untar-ing artifacts for {run_id}")
@@ -204,22 +196,15 @@ def retrieve_artifacts_by_release(args):
         release_tag = "nightly-release"
     # Otherwise, determine if version is nightly-release or dev-release
     else:
-        try:
-            version = Version(args.release)
-            if not version.is_devrelease and not version.is_prerelease:
-                log(f"This script requires a nightly-release or dev-release version.")
-                log("Please retrieve the correct release version from:")
-                log(
-                    "\t - https://github.com/ROCm/TheRock/releases/tag/nightly-release (nightly-release example: 6.4.0rc20250416)"
-                )
-                log(
-                    "\t - https://github.com/ROCm/TheRock/releases/tag/dev-release (dev-release example: 6.4.0.dev0+8f6cdfc0d95845f4ca5a46de59d58894972a29a9)"
-                )
-                log("Exiting...")
-                return
-            release_tag = "dev-release" if version.is_devrelease else "nightly-release"
-        except InvalidVersion:
-            log(f"Invalid release version '{args.release}'")
+        # Searching for nightly-release or dev-release format
+        nightly_regex_expression = (
+            "(\\d+\\.)?(\\d+\\.)?(\\*|\\d+)rc(\\d{4})(\\d{2})(\\d{2})"
+        )
+        dev_regex_expression = "(\\d+\\.)?(\\d+\\.)?(\\*|\\d+).dev0+"
+        nightly_release = re.search(nightly_regex_expression, args.release) != None
+        dev_release = re.search(dev_regex_expression, args.release) != None
+        if not nightly_release and not dev_release:
+            log("This script requires a nightly-release or dev-release version.")
             log("Please retrieve the correct release version from:")
             log(
                 "\t - https://github.com/ROCm/TheRock/releases/tag/nightly-release (nightly-release example: 6.4.0rc20250416)"
@@ -229,6 +214,8 @@ def retrieve_artifacts_by_release(args):
             )
             log("Exiting...")
             return
+
+        release_tag = "nightly-release" if nightly_release else "dev-release"
     release_version = args.release
 
     log(f"Retrieving artifacts for release tag {release_tag}")
@@ -268,7 +255,7 @@ def retrieve_artifacts_by_input_dir(args):
 
 
 def run(args):
-    log("### Provisioning TheRock 🪨 ###")
+    log("### Installing TheRock using artifacts 🪨 ###")
     _create_output_directory(args)
     if args.run_id:
         retrieve_artifacts_by_run_id(args)
@@ -292,19 +279,71 @@ def main(argv):
         "--amdgpu-family",
         type=str,
         default="gfx94X-dcgpu",
-        help="AMD GPU family to provision (please refer to this: https://github.com/ROCm/TheRock/blob/59c324a759e8ccdfe5a56e0ebe72a13ffbc04c1f/cmake/therock_amdgpu_targets.cmake#L44-L81 for family choices)",
+        help="AMD GPU family to install (please refer to this: https://github.com/ROCm/TheRock/blob/59c324a759e8ccdfe5a56e0ebe72a13ffbc04c1f/cmake/therock_amdgpu_targets.cmake#L44-L81 for family choices)",
     )
 
     # This mutually exclusive group will ensure that only one argument is present
     group = parser.add_mutually_exclusive_group(required=True)
-    group.add_argument(
-        "--run-id", type=str, help="GitHub run ID of TheRock to provision"
-    )
+    group.add_argument("--run-id", type=str, help="GitHub run ID of TheRock to install")
 
     group.add_argument(
         "--release",
         type=str,
-        help="Github release version of TheRock to provision, from the nightly-release (X.Y.ZrcYYYYMMDD) or dev-release (X.Y.Z.dev0+{hash})",
+        help="Github release version of TheRock to install, from the nightly-release (X.Y.ZrcYYYYMMDD) or dev-release (X.Y.Z.dev0+{hash})",
+    )
+
+    artifacts_group = parser.add_argument_group("artifacts_group")
+    artifacts_group.add_argument(
+        "--blas",
+        default=False,
+        help="Include 'blas' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--fft",
+        default=False,
+        help="Include 'fft' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--miopen",
+        default=False,
+        help="Include 'miopen' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--prim",
+        default=False,
+        help="Include 'prim' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--rand",
+        default=False,
+        help="Include 'rand' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--rccl",
+        default=False,
+        help="Include 'rccl' artifacts",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--tests",
+        default=False,
+        help="Include all test artifacts for enabled libraries",
+        action=argparse.BooleanOptionalAction,
+    )
+
+    artifacts_group.add_argument(
+        "--base-only", help="Include only base artifacts", action="store_true"
     )
 
     group.add_argument(